fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@25632 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSTransportable.cpp
@@ -427,7 +427,7 @@ MSTransportable::appendStage(Stage* stage, int next) {
     if (next < 0) {
         myPlan->push_back(stage);
     } else {
-        if (stepIndex + next > myPlan->size()) {
+        if (stepIndex + next > (int)myPlan->size()) {
             throw ProcessError("invalid index '" + toString(next) + "' for inserting new stage into plan of '" + getID() + "'");
         }
         myPlan->insert(myPlan->begin() + stepIndex + next, stage);
